WIP: Started refactoring
