+++ src/main/scala/io/github/edefritz/client/Tile38Client.scala
@@ -1,9 +1,11 @@
 package io.github.edefritz.client
 
-import io.github.edefritz.commands.{Bounds, Drop, FSet, Get, Output, Scan}
-import io.github.edefritz.model.RespType
+import io.circe.parser
+import io.github.edefritz.client.Tile38Client.Codec
 import io.github.edefritz.commands
-import io.github.edefritz.model.{JsonType, OutputType, RespType}
+import io.github.edefritz.commands._
+import io.github.edefritz.errors.{Tile38Error, Tile38GenericError, Tile38IdNotFoundError, Tile38KeyNotFoundError, Tile38ResponseDecodingError}
+import io.github.edefritz.model.{JsonType, OutputType, RespType, Tile38Response}
 import io.lettuce.core.RedisClient
 import io.lettuce.core.api.StatefulRedisConnection
 import io.lettuce.core.api.async.RedisAsyncCommands
@@ -14,14 +16,12 @@ import io.lettuce.core.protocol.{CommandArgs, ProtocolKeyword}
 
 import scala.concurrent.Future
 import scala.jdk.FutureConverters.CompletionStageOps
+import scala.util.{Try, Using}
 
 class Tile38Client(connectionString: String) {
-  lazy val codec: StringCodec = StringCodec.UTF8;
-  lazy val client: RedisClient = RedisClient.create(connectionString)
-  lazy val connection: StatefulRedisConnection[String, String] =
-    client.connect(codec)
-  lazy val sync: RedisCommands[String, String] = connection.sync()
-  lazy val async: RedisAsyncCommands[String, String] = connection.async()
+  private lazy val client: RedisClient = RedisClient.create(connectionString)
+
+
   // By default a lettuce connection would return RESP
   private var format: OutputType = RespType
 
@@ -33,57 +33,75 @@ class Tile38Client(connectionString: String) {
     }
   }
 
-  def get(key: String, id: String): Get = {
-    forceJson()
-    Get(key, id)(this)
+  private def argsSeqToRedisArgs(seq: Seq[Any]): CommandArgs[String, String] = {
+    val codec = StringCodec.UTF8
+    val redisArgs = new CommandArgs(codec)
+
+    seq.foreach {
+      case s: String => redisArgs.add(s)
+      case d: Double => redisArgs.add(d)
+      case i: Int    => redisArgs.add(i)
+    }
+    redisArgs
   }
 
-  def set(key: String, id: String): commands.Set = {
-    forceJson()
-    commands.Set(key, id)(this)
+  def argsSeqToRedisArgs(seq: Seq[Any]): CommandArgs[String, String] = {
+    val codec = StringCodec.UTF8
+    val redisArgs = new CommandArgs(codec)
+
+    seq.foreach {
+      case s: String => redisArgs.add(s)
+      case d: Double => redisArgs.add(d)
+      case i: Int    => redisArgs.add(i)
+    }
+    redisArgs
+  }
+  // TODO: Include this in the trait decoder later
+  private def decodeTile38Error(response: String): Tile38Error = {
+    parser.decode[Tile38GenericError](response) match {
+      case Right(genericError: Tile38GenericError) => {
+
+        genericError.err match {
+          case "key not found" =>
+            Tile38KeyNotFoundError(
+              genericError.ok,
+              genericError.err,
+              genericError.elapsed
+            )
+          case "id not found" =>
+            Tile38IdNotFoundError(
+              genericError.ok,
+              genericError.err,
+              genericError.elapsed
+            )
+          case error: String =>
+            Tile38ResponseDecodingError(s"unexpected error message: $error")
+        }
+      }
+      case Left(error) => Tile38ResponseDecodingError(error.toString)
+    }
   }
 
-  def scan(key: String): Scan = {
-    forceJson()
-    Scan(key)(this)
+  def execSync(command: Tile38Command): Try[Tile38Response] = {
+    Using(client.connect(Codec)) { client =>
+      client.sync().dispatch()
+    }
   }
 
-  def bounds(key: String): Bounds = {
-    forceJson()
-    Bounds(key)(this)
+
+  def execAsync(command: Tile38Command): Future[Tile38Response] = {
+
   }
 
-  /**
-    * Remove all objects from specified key.
-    *
-    *
-    */
-  def drop(key: String): Drop = {
-    forceJson()
-    Drop(key)(this)
-  }
-
-  def fset(key: String, id: String, fields: Map[String, Double]): FSet = {
-    forceJson()
-    FSet(key, id, fields)(this)
-  }
-
-  def dispatch(
-      commandType: ProtocolKeyword,
-      args: CommandArgs[String, String]
-  ): String = {
-    sync.dispatch(commandType, new ValueOutput(codec), args)
-  }
-
-  def dispatchAsync(
+  def exec(
       commandType: ProtocolKeyword,
       args: CommandArgs[String, String]
   ): Future[String] = {
-    async.dispatch(commandType, new ValueOutput(codec), args).asScala
+    async.dispatch(commandType, new ValueOutput(Codec), args).asScala
   }
 
-  def close() = {
-    connection.close()
-  }
+}
 
+object Tile38Client {
+  final val Codec: StringCodec = StringCodec.UTF8;
 }
