+++ src/main/scala/io/github/edefritz/model/Tile38Response.scala
@@ -1,0 +1,7 @@
+package io.github.edefritz.model
+
+trait Tile38Response {
+  val ok: Boolean
+  // TODO: we could return a proper type for duration
+  val elapsed: String
+}
